--- BitsoundReceiver.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module BitsoundReceiver {
-  umbrella header "BitsoundReceiver.h"
-
-  export *
-  module * { export * }
-}
