+++ BitsoundReceiver.framework/Resources
@@ -1,0 +1,1 @@
+Versions/Current/Resources
